+++ src/threads/threadManager.js
@@ -18,8 +18,9 @@ class ThreadWorker extends EventEmitter {
 	constructor (files, data) {
 		super();
 
-		if (Array.is(files)) {
-			files = files.map(fp => getLoadPath(fp));
+		if (!files) files = [];
+		else if (Array.is(files)) {
+			files = files.map(fp => global.getLoadPath(fp));
 		}
 		else {
 			files = global.getLoadPath(files);
@@ -30,7 +31,6 @@ class ThreadWorker extends EventEmitter {
 				data: data
 			}
 		});
-		this.id = this._worker.threadId;
 		this.tasks = new Map();
 		this.stat = ThreadWorker.Stat.IDLE;
 
@@ -45,6 +45,22 @@ class ThreadWorker extends EventEmitter {
 		});
 		this.on('suicide', this.suicide);
 	}
+	load (files) {
+		if (this.stat === ThreadWorker.Stat.DEAD) return;
+		if (Array.is(files)) {
+			files = files.map(fp => global.getLoadPath(fp));
+		}
+		else {
+			files = global.getLoadPath(files);
+		}
+		this._worker.postMessage({
+			event: 'loadfile',
+			needReply: false,
+			data: files,
+			postAt: Date.now()
+		});
+		return this;
+	}
 	send (msg) {
 		if (this.stat === ThreadWorker.Stat.DEAD) return;
 		this._worker.postMessage({
@@ -53,16 +69,22 @@ class ThreadWorker extends EventEmitter {
 			data: msg,
 			postAt: Date.now()
 		});
+		return this;
 	}
 	request (event, data, callback) {
-		if (this.stat === ThreadWorker.Stat.DEAD) return;
-		this.stat = ThreadWorker.Stat.BUSY;
-
-		var n = Date.now();
-		var eventTag = event + ':' + n;
-		this.tasks.set(eventTag, true);
-
 		return new Promise((res, rej) => {
+			if (this.stat === ThreadWorker.Stat.DEAD) {
+				if (!!callback) callback(null, 'Thread is dead.');
+				rej('Thread is dead.');
+				return;
+			}
+
+			this.stat = ThreadWorker.Stat.BUSY;
+
+			var n = Date.now();
+			var eventTag = event + ':' + n;
+			this.tasks.set(eventTag, true);
+
 			this._worker.postMessage({
 				event,
 				needReply: true,
@@ -70,23 +92,76 @@ class ThreadWorker extends EventEmitter {
 				postAt: n
 			});
 			this.once('reply:' + eventTag, (data, event) => {
-				var msg = data.data;
 				this.tasks.delete(eventTag);
 				if (this.tasks.size === 0) this.stat = ThreadWorker.Stat.IDLE;
-				if (!!callback) callback(msg);
-				res(msg);
+				if (!!callback) callback(data, null);
+				res(data);
+			});
+		});
+	}
+	evaluate (fn, data, callback) {
+		if (this.stat === ThreadWorker.Stat.DEAD) return;
+		this.stat = ThreadWorker.Stat.BUSY;
+
+		var n = Date.now();
+		var eventTag = 'evaluate:' + n;
+		this.tasks.set(eventTag, true);
+
+		return new Promise((res, rej) => {
+			this._worker.postMessage({
+				event: 'evaluate',
+				needReply: true,
+				data: {
+					fn: fn.toString(),
+					data
+				},
+				postAt: n
+			});
+			this.once('reply:' + eventTag, (data, event) => {
+				this.tasks.delete(eventTag);
+				if (this.tasks.size === 0) this.stat = ThreadWorker.Stat.IDLE;
+				if (!!data.err) {
+					if (!!callback) callback(null, data.err);
+					rej(data.err);
+				}
+				else {
+					if (!!callback) callback(data.result, null);
+					res(data.result);
+				}
 			});
 		});
 	}
 	get count () {
 		return this.tasks.size;
 	}
+	get id () {
+		if (this.stat === ThreadWorker.Stat.DEAD) return -1;
+		return this._worker.threadId;
+	}
 	suicide () {
 		this.stat = ThreadWorker.Stat.DEAD;
 		this._worker.terminate();
+		this._worker = null;
+		this.tasks = null;
+		for (let key in this._events) {
+			let cbs = this._events[key];
+			if (Function.is(cbs)) this.removeListener(key, cbs);
+			else cbs.forEach(cb => this.removeListener(key, cb));
+		}
 	}
 }
 ThreadWorker.Stat = Symbol.set(['IDLE', 'BUSY', 'DEAD']);
+
+var pool = null;
+
+const choiseThread = () => {
+	pool.sort((ta, tb) => {
+		if (ta.stat === ThreadWorker.Stat.IDLE && tb.stat === ThreadWorker.Stat.BUSY) return -1;
+		if (tb.stat === ThreadWorker.Stat.IDLE && ta.stat === ThreadWorker.Stat.BUSY) return 1;
+		return ta.count - tb.count;
+	});
+	return pool[0];
+};
 
 const TM = {
 	// 根据 filenames 批量载入运行程序，并传入初始参数 data
@@ -109,7 +184,78 @@ const TM = {
 			if (!err) res(data);
 			else rej(err);
 		});
-	})
+	}),
+	Pool: {
+		size: require('os').cpus().length,
+		create: (size=TM.Pool.size) => {
+			if (!!pool) return TM.Pool;
+			if (isNaN(size) || size <= 0) size = TM.Pool.size;
+			pool = Array.generate(size, () => new ThreadWorker());
+			return TM.Pool;
+		},
+		load: files => {
+			pool.forEach(th => th.load(files));
+			return TM.Pool;
+		},
+		request: (event, data, callback) => {
+			return new Promise(async (res, rej) => {
+				var th = choiseThread();
+				var result;
+				try {
+					result = await th.request(event, data);
+				}
+				catch (err) {
+					if (!!callback) callback(null, err);
+					rej(err);
+				}
+				if (!!callback) callback(result, null);
+				res(result);
+			});
+		},
+		requestAll: (event, data, callback) => {
+			return new Promise((res, rej) => {
+				var result = [], tasks = pool.length;;
+				pool.forEach(async (th, i) => {
+					var r;
+					try {
+						r = await th.request(event, data);
+						result[i] = [r, null];
+					}
+					catch (err) {
+						result[i] = [null, err];
+					}
+					tasks --;
+					if (tasks === 0) {
+						if (!!callback) callback(result);
+						res(result);
+					}
+				});
+			});
+		},
+		evaluate: (fn, data, callback) => {
+			return new Promise(async (res, rej) => {
+				var th = choiseThread();
+				var result;
+				try {
+					result = await th.evaluate(fn, data);
+				}
+				catch (err) {
+					if (!!callback) callback(null, err);
+					rej(err);
+				}
+				if (!!callback) callback(result, null);
+				res(result);
+			});
+		},
+		tasks: () => {
+			var t = 0;
+			pool.forEach(th => t += th.count);
+			return t;
+		},
+		killAll: () => {
+			pool.forEach(th => th.suicide());
+		}
+	}
 };
 
 module.exports = TM;